refactor: FeedCard refactored
