+++ src/domain/model/Commentable.kt
@@ -1,0 +1,6 @@
+package domain.model
+
+interface Commentable {
+    var commentCount: Int
+    fun onComment(comment: String)
+}
