+++ src/domain/model/FeedItem.kt
@@ -6,13 +6,3 @@ sealed class FeedItem(
     open val userName: String,
     open val timestamp: Long
 )
-
-interface Likeable {
-    var likeCount: Int
-    fun onLike()
-}
-
-interface Commentable {
-    var commentCount: Int
-    fun onComment(comment: String)
-}
