+++ src/domain/model/Likeable.kt
@@ -1,0 +1,6 @@
+package domain.model
+
+interface Likeable {
+    var likeCount: Int
+    fun onLike()
+}
